IDE-2151 moved all workspace wizards to gradle plugins
