+++ gradle/plugins/com.liferay.ide.gradle.core/META-INF/MANIFEST.MF
@@ -21,6 +21,7 @@ Require-Bundle: org.eclipse.core.runtime,
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ActivationPolicy: lazy
 Export-Package: com.liferay.ide.gradle.core,
+ com.liferay.ide.gradle.core.workspace,
  com.liferay.ide.gradle.toolingapi.custom
 Bundle-ClassPath: .,
  repo/com/liferay/ide/gradle/toolingapi/model/1.1.1/model-1.1.1.jar
