+++ gradle/plugins/com.liferay.ide.gradle.core/src/com/liferay/ide/gradle/core/workspace/LiferayWorkspaceUtil.java
@@ -13,16 +13,18 @@
  *
  *******************************************************************************/
 
-package com.liferay.ide.project.core.workspace;
+package com.liferay.ide.gradle.core.workspace;
 
 import com.liferay.ide.core.util.CoreUtil;
 import com.liferay.ide.core.util.FileUtil;
 import com.liferay.ide.project.core.ProjectCore;
+import com.liferay.ide.project.core.util.ProjectImportUtil;
 
 import java.io.File;
 
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IStatus;
 
 /**
  * @author Andy Wu
@@ -34,6 +36,21 @@ public class LiferayWorkspaceUtil
 
     public static String hasLiferayWorkspaceMsg =
         "A Liferay Workspace project already exists in this Eclipse instance.";
+
+    public static IStatus validateWorkspacePath(final String currentPath)
+    {
+        IStatus retVal = ProjectImportUtil.validatePath( currentPath );
+
+        if( retVal.isOK() )
+        {
+            if( !LiferayWorkspaceUtil.isValidWorkspaceLocation( currentPath ) )
+            {
+                retVal = ProjectCore.createErrorStatus( "Invalid Liferay Workspace" );
+            }
+        }
+
+        return retVal;
+    }
 
     public static void clearWorkspace( String location )
     {
