+++ gradle/plugins/com.liferay.ide.gradle.core/src/com/liferay/ide/gradle/core/workspace/NewLiferayWorkspaceOp.java
@@ -12,7 +12,7 @@
  * details.
  *
  *******************************************************************************/
-package com.liferay.ide.project.core.workspace;
+package com.liferay.ide.gradle.core.workspace;
 
 import org.eclipse.sapphire.ElementType;
 import org.eclipse.sapphire.ExecutableElement;
