+++ gradle/plugins/com.liferay.ide.gradle.core/src/com/liferay/ide/gradle/core/workspace/ServerNameDefaultValueService.java
@@ -13,7 +13,7 @@
  *
  *******************************************************************************/
 
-package com.liferay.ide.project.core.workspace;
+package com.liferay.ide.gradle.core.workspace;
 
 import com.liferay.ide.core.util.StringPool;
