+++ gradle/plugins/com.liferay.ide.gradle.ui/META-INF/MANIFEST.MF
@@ -13,7 +13,14 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.debug.core,
  org.eclipse.buildship.core,
  org.gradle.toolingapi,
- com.gradleware.tooling.model
+ com.gradleware.tooling.model,
+ org.eclipse.sapphire.ui;bundle-version="8.2.1",
+ com.liferay.ide.gradle.core,
+ org.eclipse.ui.ide,
+ org.eclipse.wst.web,
+ org.eclipse.e4.core.contexts,
+ org.eclipse.e4.core.commands,
+ org.eclipse.core.expressions
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
