+++ gradle/plugins/com.liferay.ide.gradle.ui/src/com/liferay/ide/gradle/ui/workspace/NewLiferayWorkspaceWizard.png
