+++ tools/plugins/com.liferay.ide.project.core/META-INF/MANIFEST.MF
@@ -49,8 +49,7 @@ Export-Package: com.liferay.ide.project.core,
  com.liferay.ide.project.core.model,
  com.liferay.ide.project.core.modules,
  com.liferay.ide.project.core.upgrade,
- com.liferay.ide.project.core.util,
- com.liferay.ide.project.core.workspace
+ com.liferay.ide.project.core.util
 Bundle-ClassPath: lib/jackson-core-asl-1.9.13.jar,
  lib/jackson-mapper-asl-1.9.13.jar,
  .
